Refactor: extract applyMessageOptions and rename Objects to DTO
- Extracted duplicate sender/quickReply logic to applyMessageOptions()
- Renamed Phine\Objects -> Phine\DTO for clarity
- Used arrow functions for cleaner array_map callbacks
